About Jabaroo
Our Jabaroo NFT project journey started with displaying the assets we designed with the NFT creator account in an NFT Market. Based on this, we determined the directions that we can develop in this sector and aimed to deal with 2 different issues in total.
We will create Jabaroo NFT Market and aim to integrate with leading NFT markets in the marketplace.
In addition to the NFT Market we have established, a Physical Market will be created where we will sell and market physical NFTs. We are very confident in this idea, which we think is lacking in the market at the moment, and we aim to be the world's first physical e-commerce platform that sells with smart contracts.
Creating the Jabaroo NFT Marketplace and Provide Integration with leading NFT Markets

We aim to develop a Jabaroo NFT Marketplace with the Jabaroo NFT project. By creating an account here, users will be able to generate, buy or sell NFT and will not need any coding skills to do so.
In this market, which we will create by designing a user-friendly interface, users will be able to start trading after opening their accounts as collectors or creators. However, users' accounts will need to be verified for their work to stand out. With the verification system we will develop here, we will ensure that users can trade securely. With this market, we aim to eliminate the slow user authentication process experienced in NFT markets available today. Feedback will also be given to users who do not meet the conditions for the verification of the account, as to why their account was not verified.
We are aware that the NFT craze, which has become widespread today, has taken place in many different markets. We think that it is very inefficient for users to open an account in each market in order to reach NFT projects in different markets. As a solution to this problem, we aim to integrate the leading NFT markets in the marketplace with the Jabaroo NFT market. Thus, users will be able to make transactions through other markets that we integrate with without the need to open a new account. The accounts they have opened in the Jabaroo NFT market will be enough. In order to proceed to this step, users will need to be verified on the Jabaroo NFT Market, and then all integrated markets will be open to transactions. Thus, the requirement of opening different accounts and user verification for each market will be eliminated. It will save time and create an integrated NFT Marketplace ecosystem.
For sales on Jabaroo NFT Market, Jabaroo will only receive 2.5% commission from the buyer and seller on the sales made.
Creating a Physical NFT Marketplace

The NFT market has shown to all of us that it is an innovative and open to improvement area. Thanks to NFT, billions of works created around the world meet with their buyers and gain the value they deserve. As a team, we are dealing with the physical NFT Market project that will increase this value even more.
We aim to create a Physical NFT Marketplace as a different branch of the NFT Market that we will develop under the Jabaroo NFT Project. We will establish a structure where digitally produced assets in the current NFT market can be physically handled and objectified.
Thanks to the Physical NFT Market;
The works designed as digital assets will be presented to the buyers by objectifying them with 3D printers.
Creators will be able to make their own designs in 3D if they wish. They will be able to make their sales by uploading the images of the 3D work to the market. In this scenario, Jabaroo will only receive a 3% commission from the buyer and seller on the sale made.
Creators who have designed digitally but do not have the equipment to make it 3D will be able to share their designs with Jabaroo Physical NFT Market Team. Jabaroo Team will make these designs to 3D version and provide logistics if they are sold. In this scenario, Jabaroo Team will make a price offer to Creator, according to the difficulty of transforming the digital work into 3D. If Creator accepts, Jabaroo Team will make the work 3D and share the photos with Creator. If the creator requests, the created work sample will be sent as a one-time. The photos of the product will be shared by Creator on its own page. If the asset is sold, Creator will send an order to Jabaroo and the logistics of the product will be done by Jabaroo. In this case, Jabaroo Team will receive 3% commission from the buyer and seller on the sale made in addition to the price offer it has given.
Works designed as digital assets can be printed on T-shirts using printing machines and presented to the buyers.
Creatorlar isterlerse kendi tasarımlarını kendileri T-shirtlere baskı Creators will be able to print their own designs on T-shirts if they wish, and they will be able to make their sales by uploading the images of the printed assets to the market. In this scenario, Jabaroo will only receive a 3% commission from the buyer and seller on the sale made.
Creators who have designed digitally but do not have the equipment to print on T-shirts will be able to share their designs with the Jabaroo Physical NFT Market Team. Jabaroo Team will print these designs on T-shirts and provide logistics if they are sold. In this scenario, Jabaroo Team will make a price offer to Creator, depending on the difficulty of converting the digital work to print. If Creator accepts, Jabaroo Team will print the work on the T-shirt and share the photographs with Creator. If the creator requests, the created work sample will be sent as a one-time. The photos of the product will be shared by Creator on its own page. If it is sold, Creator will send an order to Jabaroo Team and the logistics of the product will be done by Jabaroo. In this case, Jabaroo Team will receive 3% commission from the buyer and seller on the sale made in addition to the price offer it has given.

In both themes, it will be presented as a condition that the products are first created as digital works and then transformed into physical works Jabaroo Physical NFT Market will support multiple photo insertion and creators will be expected to add photos of digital and physical forms of their work.Works that are not transformed into physical works and photographed will not be listed and cannot be sold.
In order for the users to make transactions in the Physical NFT Market, it is expected that they have created their users as creators or collectors on the NFT Market and have their accounts verified.
While the Physical NFT Market is being designed, a new Smart Contract will be designed for safe trading. Smart Contracts used in the market were designed in accordance with digital works. Since digital works will be physically sold with the Physical NFT Market to be designed by Jabaroo, there is a need to design a new Smart Contract. This issue will be handled within the scope of the project and a platform will be created for users to use safely.

FOOTNOTE: Both market places will be traded with the Binance Smart Chain network. Here, transactions will be carried out quickly and with low transaction fees. Also, all data will be safe with the blockchain network.